% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactmass_calc.R
\name{calculate_monoisotope}
\alias{calculate_monoisotope}
\title{Calculate the monoisotopic mass of an elemental formula}
\usage{
calculate_monoisotope(elementcomp, ionstate = "none")
}
\arguments{
\item{elementcomp}{character string containing an elemental formula}

\item{ionstate}{character string containing the ion state of the formula in `[M+H]+` format}
}
\value{
numeric monoisotopic mass
}
\description{
Calculate the monoisotopic mass of an elemental formula
}
\examples{
calculate_monoisotope("C2H6O6", ionstate = "[M+H]+")
}
