% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exactmasschart}
\alias{exactmasschart}
\title{Exact mass data for all elements}
\format{
## `who`
A list containing three elements
\describe{
  \item{elements}{elemental notation of individual elements}
  \item{masses}{matrix of monoisotopic masses of each isotope of the elements}
  \item{abundances}{matrix of the natural abundances of each isotope of the elements}
  ...
}
}
\source{
<https://www.nist.gov/pml/atomic-weights-and-isotopic-compositions-relative-atomic-masses>
}
\usage{
exactmasschart
}
\description{
A list of exact masses isotopic abundances of all elements
}
\keyword{datasets}
