% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactmass_calc.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract individual elements from a formula string}
\usage{
extract_elements(composition.str)
}
\arguments{
\item{composition.str}{character string containing an elemental formula}
}
\value{
list object containing elements and counts of individual elements
}
\description{
Extract individual elements from a formula string
}
\examples{
extract_elements("C2H6O6")
}
