% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{isotopic_distribution}
\alias{isotopic_distribution}
\title{Generate isotopic distribution mass spectrum}
\usage{
isotopic_distribution(
  elementalformula,
  max.dist = 3,
  min.int = 0.001,
  charge = "neutral"
)
}
\arguments{
\item{elementalformula}{character string of chemical formula}

\item{max.dist}{numeric maximum absolute distance from the lowest m/z value to include in mass spectrum}

\item{min.int}{numeric minimum relative intensity to include in mass spectrum (between 0 and 1)}

\item{charge}{character vector of the charge state of the mass spectrum ("positive", "negative" or "neutral")}
}
\value{
table containing m/z values (mz) and intensity values (int) of a predicted isotopic pattern
}
\description{
Generate isotopic distribution mass spectrum
}
\examples{
isotopic_distribution("C8F12O2", charge = "negative")
}
