% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition_prediction.R
\name{predictcomposition}
\alias{predictcomposition}
\title{Predict possible elemental compositions for a specific mass}
\usage{
predictcomposition(
  mass,
  error,
  min.error,
  elements = c("C", "H", "N", "O", "F", "S"),
  counts = list(0:39, 0:72, 0:20, 0:20, 0:20, 0:10)
)
}
\arguments{
\item{mass}{numeric mass to predict elemental composition (in Da)}

\item{error}{numeric relative error (in ppm)}

\item{min.error}{numeric minimum absolute error  (in Da)}

\item{elements}{list of characters containing elements to include in calculation}

\item{counts}{list of integer vectors of count of elements to include in calculation}
}
\value{
table containing elemental combinations, calculated mass, ppm error, and double bond equivalents (DBE)
}
\description{
Predict possible elemental compositions for a specific mass
}
\examples{
predictcomposition(498.9302, error = 5, min.error = 0.002, elements = c("C", "H", "O", "N", "F", "S"), counts = list(0:39, 0:72, 0:20, 0:20, 0:20, 0:10))
}
